Discount

+++ program52.rb
@@ -1,7 +1,7 @@
 #calculate discount
 
 print "Enter price : "
-price = gets.chomp.to_i
+price = gets.chomp.to_f
 
 print "Enter discount : "
 discount = gets.chomp.to_f
